+++ src/pconn.h
@@ -59,14 +59,16 @@ public:
 
     void clearHandlers(const Comm::ConnectionPointer &conn);
 
+    // TODO: Upgrade to return size_t
     int count() const { return size_; }
+
     void closeN(size_t count);
 
     // IndependentRunner API
     void endingShutdown() override;
 private:
     bool isAvailable(int i) const;
-    bool removeAt(int index);
+    bool removeAt(size_t index);
     int findIndexOf(const Comm::ConnectionPointer &conn) const;
     void findAndClose(const Comm::ConnectionPointer &conn);
     static IOCB Read;
@@ -81,9 +83,9 @@ private:
     Comm::ConnectionPointer *theList_;
 
     /// Number of entries theList can currently hold without re-allocating (capacity).
-    int capacity_;
+    size_t capacity_;
     ///< Number of in-use entries in theList
-    int size_;
+    size_t size_;
 
     /** The pool containing this sub-list.
      * The parent performs all stats accounting, and
